+++ flash/tabular/regression/model.py
@@ -17,7 +17,7 @@ import torch
 from torch.nn import functional as F
 
 from flash.core.data.data_source import DefaultDataKeys
-from flash.core.Regression import RegressionTask
+from flash.core.regression import RegressionTask
 from flash.core.utilities.imports import _TABULAR_AVAILABLE
 from flash.core.utilities.types import LR_SCHEDULER_TYPE, METRICS_TYPE, OPTIMIZER_TYPE, SERIALIZER_TYPE
